requirements.txt
tqdm==4.64.0
sklearn==0.0.post1
pandas==1.4.2
numpy==1.21.5
matplotlib==3.5.1
seaborn==0.11.2
imblearn==0.0
joblib==1.2.0
json5==0.9.6
PyYAML==6.0
requests==2.27.1
fastapi==0.88.0
pydantic==1.10.2
uvicorn==0.20.0
streamlit==1.16.0
Pillow==9.0.1